requirements.txt
# Core dependencies
groq>=0.31.0
python-dotenv>=1.0.0

# Additional professional dependencies
pathlib2>=2.3.7
typing-extensions>=4.8.0

# API dependencies
fastapi>=0.116.1
uvicorn[standard]>=0.21.1
python-multipart>=0.0.6

# Testing dependencies
requests>=2.31.0

# Frontend dependencies
streamlit>=1.27.0
